+++ CotEditor/Sources/DotView.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2015-2018 1024jp
+//  © 2015-2022 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -23,14 +23,15 @@
 //  limitations under the License.
 //
 
-import Cocoa
+import AppKit
 
 @IBDesignable
 final class DotView: NSView {
     
     // MARK: Inspectable Properties
     
-    @IBInspectable var color: NSColor = .labelColor
+    @Invalidating(.display) @IBInspectable var color: NSColor = .labelColor
+    @Invalidating(.display) @IBInspectable var dotLength: CGFloat = 4
     
     
     
@@ -39,10 +40,18 @@ final class DotView: NSView {
     
     override func draw(_ dirtyRect: NSRect) {
         
+        assert(self.dotLength <= self.bounds.width)
+        assert(self.dotLength <= self.bounds.height)
+        
+        let rect = NSRect(x: (self.bounds.width - self.dotLength) / 2,
+                          y: (self.bounds.height - self.dotLength) / 2,
+                          width: self.dotLength,
+                          height: self.dotLength)
+        
         NSGraphicsContext.saveGraphicsState()
         
         self.color.setFill()
-        NSBezierPath(ovalIn: self.bounds).fill()
+        NSBezierPath(ovalIn: rect).fill()
         
         NSGraphicsContext.restoreGraphicsState()
     }